fix: matsubara precision factor is now (should be) double

+++ DiFfRG/include/DiFfRG/common/quadrature/quadrature_provider.hh
@@ -48,7 +48,7 @@ namespace DiFfRG
       void set_vacuum_quad_size(const int size);
       void set_min_matsubara_size(const int value);
       void set_max_matsubara_size(const int value);
-      void set_matsubara_precision_factor(const int value);
+      void set_matsubara_precision_factor(const double value);
 
     private:
       MatsubaraQuadrature<double> &get_matsubara_quadrature_d(const double T, const double E);
@@ -71,7 +71,7 @@ namespace DiFfRG
 
       int verbosity = 0;
       int vacuum_quad_size = 64;
-      int matsubara_precision_factor = 0;
+      double matsubara_precision_factor = 0;
       int min_matsubara_size = 32;
       int max_matsubara_size = 128;
